moved configurations to auth-options
