+++ src/management-system-v2/app/api/auth/[...nextauth]/auth-options.ts
@@ -1,11 +1,11 @@
 import { AuthOptions } from 'next-auth';
 import { User } from '@/types/next-auth';
 import { randomUUID } from 'crypto';
-import { getRoles } from '@/lib/data/legacy/iam/roles';
-import { addRoleMappings } from '@/lib/data/legacy/iam/role-mappings';
+import Auth0Provider from 'next-auth/providers/auth0';
+import CredentialsProvider from 'next-auth/providers/credentials';
 import { addUser, usersMetaObject } from '@/lib/data/legacy/iam/users';
 
-export const nextAuthOptions: AuthOptions = {
+const nextAuthOptions: AuthOptions = {
   secret: process.env.NEXTAUTH_SECRET,
   providers: [],
   callbacks: {
@@ -44,3 +44,63 @@ export const nextAuthOptions: AuthOptions = {
     },
   },
 };
+
+if (process.env.USE_AUTH0) {
+  nextAuthOptions.providers.push(
+    Auth0Provider({
+      clientId: process.env.AUTH0_CLIENT_ID as string,
+      clientSecret: process.env.AUTH0_CLIENT_SECRET as string,
+      issuer: process.env.AUTH0_ISSUER,
+      authorization: {
+        params: {
+          scope: process.env.AUTH0_SCOPE,
+        },
+      },
+      profile(profile) {
+        return {
+          id: `auth0:${profile.sub}`,
+          email: profile.email,
+          image: profile.picture,
+          firstName: profile.given_name,
+          lastName: profile.family_name,
+          username: profile.preferred_username,
+        };
+      },
+    }),
+  );
+}
+
+if (process.env.NODE_ENV === 'development') {
+  const developmentUsers = [
+    {
+      username: 'johndoe',
+      firstName: 'John',
+      lastName: 'Doe',
+      email: 'johndoe@proceed-labs.org',
+      id: 'development-id|johndoe',
+    },
+    {
+      username: 'admin',
+      firstName: 'Admin',
+      lastName: 'Admin',
+      email: 'admin@proceed-labs.org',
+      id: 'development-id|admin',
+    },
+  ];
+
+  nextAuthOptions.providers.push(
+    CredentialsProvider({
+      name: 'Development Users',
+      credentials: {
+        username: { label: 'Username', type: 'text', placeholder: 'johndoe | admin' },
+      },
+      async authorize(credentials) {
+        const user = developmentUsers.find((user) => user.username === credentials?.username);
+
+        return user || null;
+      },
+    }),
+  );
+}
+
+export default nextAuthOptions;
